requirements.txt
# Alchemy API Toolkit Requirements
requests>=2.28.1
web3>=6.0.0
python-dotenv>=0.21.0
aiohttp>=3.8.3
asyncio>=3.4.3
